// packages/components/DataTable/models/props/DataGridProps.ts
export interface DataGridProcessedProps<R extends GridValidRowModel = any>
    extends DataGridPropsWithDefaultValues,
        DataGridPropsWithComplexDefaultValueAfterProcessing,
        Omit<DataGridPropsWithoutDefaultValue<R>, 'componentsProps'> {}
// ...
export interface DataGridPropsWithComplexDefaultValueAfterProcessing {
    slots: UncapitalizeObjectKeys<GridSlotsComponent>;
    localeText: GridLocaleText;
}